+++ plugins/destination/duckdb/main.go
@@ -21,6 +21,7 @@ func main() {
 		client.New,
 		plugin.WithBuildTargets([]plugin.BuildTarget{
 			{OS: plugin.GoOSLinux, Arch: plugin.GoArchAmd64, CGO: true},
+			{OS: plugin.GoOSLinux, Arch: plugin.GoArchArm64, CGO: true},
 			{OS: plugin.GoOSDarwin, Arch: plugin.GoArchAmd64, CGO: true},
 			{OS: plugin.GoOSDarwin, Arch: plugin.GoArchArm64, CGO: true},
 		}),