+++ plugins/destination/mysql/client/migrate.go
@@ -68,6 +68,8 @@ func (*Client) canAutoMigrate(changes []schema.TableColumnChange) bool {
 			if change.Previous.PrimaryKey || change.Previous.NotNull {
 				return false
 			}
+		case schema.TableColumnChangeTypeRemoveUniqueConstraint:
+			return true
 		default:
 			return false
 		}