+++ plugins/destination/snowflake/main.go
@@ -22,6 +22,7 @@ func main() {
 		client.New,
 		plugin.WithBuildTargets([]plugin.BuildTarget{
 			{OS: plugin.GoOSLinux, Arch: plugin.GoArchAmd64},
+			{OS: plugin.GoOSLinux, Arch: plugin.GoArchArm64},
 			{OS: plugin.GoOSWindows, Arch: plugin.GoArchAmd64, CGO: true},
 			{OS: plugin.GoOSDarwin, Arch: plugin.GoArchAmd64},
 			{OS: plugin.GoOSDarwin, Arch: plugin.GoArchArm64},