+++ plugins/destination/sqlite/main.go
@@ -20,6 +20,13 @@ func main() {
 		internalPlugin.Name,
 		internalPlugin.Version,
 		client.New,
+		plugin.WithBuildTargets([]plugin.BuildTarget{
+			{OS: plugin.GoOSLinux, Arch: plugin.GoArchAmd64, CGO: true},
+			{OS: plugin.GoOSLinux, Arch: plugin.GoArchArm64, CGO: true, Env: []string{"CC=aarch64-linux-gnu-gcc"}},
+			{OS: plugin.GoOSWindows, Arch: plugin.GoArchAmd64, CGO: true},
+			{OS: plugin.GoOSDarwin, Arch: plugin.GoArchAmd64, CGO: true},
+			{OS: plugin.GoOSDarwin, Arch: plugin.GoArchArm64, CGO: true},
+		}),
 		plugin.WithKind(internalPlugin.Kind),
 		plugin.WithTeam(internalPlugin.Team),
 		plugin.WithJSONSchema(client.JSONSchema),